feat(runtime): add Unix app-bound role registry (KEL-75 T2)
Host-owned RoleRegistry supervises one primary and one app-bound
authenticated Bun child independently, so a crash/restart or foreign
token on one role cannot revoke the other.
