+++ docs/agents/learnings.md
@@ -106,5 +106,6 @@ small — it is loaded by every agent session.
 - 2026-08-19 [runtime] KEL-78 forbids inherit-from-host and child-ambient sockets from a guard grant; leaving Strict is admission failure, not an unsandboxed restart. (evidence: KEL-78, `docs/specs/kel78-strict-profile-sandbox.md`)
 - 2026-08-19 [shell] In zsh, lowercase `path` is tied to `PATH`; using `path` as a loop variable can erase command lookup mid-script. Use names like `changed_file` in hygiene/review loops. (evidence: KEL-75 review hygiene command, `docs/agents/learnings.md`)
 - 2026-08-19 [ipc] A Unix bootstrap cancellation wake socket is only a nudge; on macOS `shutdown()` on the freshly connected wake stream can race the accepting side and return `NotConnected`. Treat that as successful cancellation, not a real listener failure. (evidence: KEL-75, `crates/keld-ipc/src/bootstrap.rs` `shutdown_unblocks_accept_and_removes_owner_only_directory`)
+- 2026-08-19 [runtime] Latch bootstrap `admission_done` on deadline/failure/disconnect; otherwise a later `revoke` can overwrite the real admission error with `Disconnected`. (evidence: KEL-75 T1b review, `crates/keld-runtime/src/unix_role.rs`)
 - 2026-08-19 [runtime] Bun 1.3.14 `subprocess.kill()` called after the child's `'exit'` returns `true` and sets `subprocess.killed = true` even though a raw `process.kill(pid, 0)` at the same instant throws `ESRCH`; Node v26.7.0 returns `false`/`false`. A supervisor that branches on `kill()` to decide whether to escalate to `SIGKILL` reads success from a signal that was never delivered. Divergences in this family are parent-attributed — always probe the parent x child runtime matrix before blaming the child. (evidence: KEL-77, `cargo nextest run -p keld-runtime bun_kill_after_exit_is_the_pinned_defect`, crates/keld-runtime/fixtures/child-process/)
 - 2026-08-19 [runtime] Dropping path-NUL-length framing from the KEL-77 corpus digest still passes a one-byte fixture flip; `("ab","c")` vs `("a","bc")` is the test that fails (concatenated SHA-256 of `abc`). (evidence: KEL-77 T2 negative control, `corpus_digest_rename_plus_content_shuffle_does_not_collide`)
