+++ llms-full.txt
@@ -974,11 +974,12 @@ everywhere except `keld-wv` backends and the shm module of `keld-ipc`, where eac
   `--inspect` passthrough in dev. Every destination spawn is a fresh principal/link
   generation—not a PID, token or socket name—and old authority is revoked before a
   successor is provisioned. Current implementation has KEL-70's generic one-child
-  supervision, the fixed CLI echo link, and KEL-75 T1b's Unix
-  `keld_runtime::primary::PrimaryRoleSupervisor` for one primary role with fresh
-  restart identity and stale-token rejection. Additional role categories, virtual
-  ports, role grants, strict sandbox admission, and Windows named-pipe/DACL bootstrap
-  remain later KEL-75/KEL-78 slices.
+  supervision, the fixed CLI echo link, KEL-75 T1b's Unix authenticated role
+  coordinator (`keld_runtime::primary`), and KEL-75 T2's Unix
+  `keld_runtime::registry::RoleRegistry` for one `primary` plus one independent
+  `app-bound` role. Window-bound roles, virtual ports, role grants, strict
+  sandbox admission, and Windows named-pipe/DACL bootstrap remain later
+  KEL-75/KEL-78 slices.
 - Webview content processes: whatever the selected engine does (WKWebView WebContent,
   WebView2 helpers, WebKitGTK web process, or future CEF subprocesses if that candidate
   lands). We never fight the engine's model.
@@ -1833,14 +1834,13 @@ never a PID recovered after exit.
 `keld.config.ts` owns entry/lifecycle declaration; `keld.permissions.jsonc` owns the
 generated capability subset and any separately reviewed role-specific addition. No
 environment identity, child payload, token, PID or facade option can choose a role or
-authority. Current implementation is still one primary role, not a role family: it has
-KEL-70's generic one-child supervisor, KEL-75 T1a's Unix authenticated bootstrap
-listener, and KEL-75 T1b's Unix `keld_runtime::primary::PrimaryRoleSupervisor`. The T1b
-coordinator mints a fresh endpoint/token per restart, binds only after a valid `HELLO`,
-reports redacted host-only bootstrap rejection, and revokes before successor
-provisioning while reusing the generic supervisor for spawn/backoff/reap. It does not
-implement a multi-role registry, role-specific grants, virtual ports, strict OS
-sandboxing, or Windows named-pipe/DACL bootstrap.
+authority. Current Unix implementation is a two-slot host registry, not a complete
+role family: KEL-70's generic one-child supervisor, KEL-75 T1a's Unix authenticated
+bootstrap listener, T1b's per-role generation coordinator, and T2's
+`keld_runtime::registry::RoleRegistry` which owns one `primary` and one `app-bound`
+supervisor independently. A primary restart does not revoke or stop the app-bound
+role. It does not implement window-bound lifecycle, role-specific grants, virtual
+ports, strict OS sandboxing, or Windows named-pipe/DACL bootstrap.
 
 The ordered destination flow below is KEL-75's source of truth for spawn, port routing,
 window close and restart. KEL-78 separately owns real-OS sandbox admission proof.
@@ -1893,9 +1893,10 @@ does not obtain a raw child endpoint, mapping handle or authority to spawn a pro
 Ports are FIFO per generation, transfers are one-shot and receiver-bound, and close or
 generation revocation disconnects the peer without exposing another principal. Exact
 Electron-observable queue/start, transfer validation and close-event behavior is owned
-by pinned conformance entries—not assumed from this generic runtime contract. The
-first implementation slice is one primary role with fresh identity/restart proof; named
-roles and ports follow only after that slice passes.
+by pinned conformance entries—not assumed from this generic runtime contract. Live Unix
+slices are T1b (one authenticated role generation) and T2 (one primary plus one
+independent app-bound role in `RoleRegistry`). Window-bound roles and virtual ports
+follow only after those slices.
 
 ## 2. keld CLI: verbs and guarantees
